09_권윤형/Retriever/requirements.txt
langchain_openai
langchain_community
rank_bm25
faiss-cpu